test(interfaces/slurm): remove last event checks
