+++ tests/unit/interfaces/oci_runtime/test_oci_runtime.py
@@ -16,17 +16,71 @@
 
 from collections import defaultdict
 
+import ops
 import pytest
-from mock_charms import (
-    EXAMPLE_OCI_CONFIG,
-    OCI_RUNTIME_INTEGRATION_NAME,
-    MockOCIRunTimeProviderCharm,
-    MockOCIRunTimeRequirerCharm,
-)
 from ops import testing
 from slurmutils import OCIConfig
 
-from hpc_libs.interfaces import OCIRuntimeDisconnectedEvent, OCIRuntimeReadyEvent
+from hpc_libs.interfaces import (
+    OCIRuntimeData,
+    OCIRuntimeDisconnectedEvent,
+    OCIRuntimeProvider,
+    OCIRuntimeReadyEvent,
+    OCIRuntimeRequirer,
+    SlurmctldConnectedEvent,
+)
+
+OCI_RUNTIME_INTEGRATION_NAME = "oci-runtime"
+EXAMPLE_OCI_CONFIG = OCIConfig(
+    ignorefileconfigjson=False,
+    envexclude="^(SLURM_CONF|SLURM_CONF_SERVER)=",
+    runtimeenvexclude="^(SLURM_CONF|SLURM_CONF_SERVER)=",
+    runtimerun="apptainer exec --userns %r %@",
+    runtimekill="kill -s SIGTERM %p",
+    runtimedelete="kill -s SIGKILL %p",
+)
+
+
+class MockOCIRunTimeProviderCharm(ops.CharmBase):
+    """Mock OCI runtime provider charm."""
+
+    def __init__(self, framework: ops.Framework) -> None:
+        super().__init__(framework)
+
+        self._oci_runtime = OCIRuntimeProvider(self, OCI_RUNTIME_INTEGRATION_NAME)
+
+        framework.observe(self._oci_runtime.on.slurmctld_connected, self._on_slurmctld_connected)
+
+    def _on_slurmctld_connected(self, event: SlurmctldConnectedEvent) -> None:
+        self._oci_runtime.set_oci_runtime_data(
+            OCIRuntimeData(ociconfig=EXAMPLE_OCI_CONFIG),
+            integration_id=event.relation.id,
+        )
+
+
+class MockOCIRunTimeRequirerCharm(ops.CharmBase):
+    """Mock OCI runtime requirer charm."""
+
+    def __init__(self, framework: ops.Framework) -> None:
+        super().__init__(framework)
+
+        self._oci_runtime = OCIRuntimeRequirer(self, OCI_RUNTIME_INTEGRATION_NAME)
+
+        framework.observe(
+            self._oci_runtime.on.oci_runtime_ready,
+            self._on_oci_runtime_ready,
+        )
+        framework.observe(
+            self._oci_runtime.on.oci_runtime_disconnected,
+            self._on_oci_runtime_disconnected,
+        )
+
+    def _on_oci_runtime_ready(self, event: OCIRuntimeReadyEvent) -> None:
+        config = self._oci_runtime.get_oci_runtime_data(event.relation.id)
+        # Assume `remote_app_data` contains `oci.conf` configuration data.
+        assert config.ociconfig.dict() == EXAMPLE_OCI_CONFIG.dict()
+
+    def _on_oci_runtime_disconnected(self, event: OCIRuntimeDisconnectedEvent) -> None: ...
 
 
 @pytest.fixture(scope="function")
@@ -34,7 +88,7 @@ def provider_ctx() -> testing.Context[MockOCIRunTimeProviderCharm]:
     return testing.Context(
         MockOCIRunTimeProviderCharm,
         meta={
-            "name": "oci-runtime-provides",
+            "name": "oci-runtime-provider",
             "provides": {OCI_RUNTIME_INTEGRATION_NAME: {"interface": "slurm_oci_runtime"}},
         },
     )
@@ -45,112 +99,118 @@ def requirer_ctx() -> testing.Context[MockOCIRunTimeRequirerCharm]:
     return testing.Context(
         MockOCIRunTimeRequirerCharm,
         meta={
-            "name": "oci-runtime-requires",
+            "name": "oci-runtime-requirer",
             "requires": {OCI_RUNTIME_INTEGRATION_NAME: {"interface": "slurm_oci_runtime"}},
         },
     )
 
 
-@pytest.mark.parametrize("leader", (True, False))
-def test_slurmctld_connected_event_handler(provider_ctx, leader) -> None:
-    """Test that an OCI runtime provider correctly sets `oci.conf` data in application data."""
-    oci_runtime_integration_id = 22
-    oci_runtime_integration = testing.Relation(
-        endpoint=OCI_RUNTIME_INTEGRATION_NAME,
-        interface="oci-runtime",
-        id=oci_runtime_integration_id,
-        remote_app_name="slurmctld",
-    )
+@pytest.mark.parametrize(
+    "leader",
+    (
+        pytest.param(True, id="leader"),
+        pytest.param(False, id="not leader"),
+    ),
+)
+class TestSlurmOCIRunTimeInterface:
+    """Unit tests for the `slurm-oci-runtime` integration interface implementation."""
 
-    state = provider_ctx.run(
-        provider_ctx.on.relation_created(oci_runtime_integration),
-        testing.State(
-            leader=leader,
-            relations={oci_runtime_integration},
+    # Test provider-side of the `slurm-oci-runtime` interface.
+
+    def test_provider_slurmctld_connected_event(self, provider_ctx, leader) -> None:
+        """Test that an OCI runtime provider correctly sets `oci.conf` data in application data."""
+        oci_runtime_integration_id = 1
+        oci_runtime_integration = testing.Relation(
+            endpoint=OCI_RUNTIME_INTEGRATION_NAME,
+            interface="oci-runtime",
+            id=oci_runtime_integration_id,
+            remote_app_name="oci-runtime-requirer",
+        )
+
+        state = provider_ctx.run(
+            provider_ctx.on.relation_created(oci_runtime_integration),
+            testing.State(
+                leader=leader,
+                relations={oci_runtime_integration},
+            ),
+        )
+
+        integration = state.get_relation(oci_runtime_integration_id)
+        if leader:
+            # Verify that the leader unit has set `oci.conf` data in `local_app_data`.
+            assert "ociconfig" in integration.local_app_data
+            config = OCIConfig.from_json(integration.local_app_data["ociconfig"])
+            assert config.dict() == EXAMPLE_OCI_CONFIG.dict()
+        else:
+            # Verify that non-leader units have not set anything in `local_app_data`.
+            assert integration.local_app_data == {}
+
+    # Test requirer-side of the `slurm-oci-runtime` interface.
+
+    @pytest.mark.parametrize(
+        "ready",
+        (
+            pytest.param(True, id="ready"),
+            pytest.param(False, id="not ready"),
         ),
     )
-
-    integration = state.get_relation(oci_runtime_integration_id)
-    if leader:
-        # Verify that the leader unit has set `oci.conf` data in `local_app_data`.
-        assert "ociconfig" in integration.local_app_data
-        config = OCIConfig.from_json(integration.local_app_data["ociconfig"])
-        assert config.dict() == EXAMPLE_OCI_CONFIG.dict()
-    else:
-        # Verify that non-leader units have not set anything in `local_app_data`.
-        assert integration.local_app_data == {}
-
-
-@pytest.mark.parametrize("remote_app_data", ({}, {"ociconfig": EXAMPLE_OCI_CONFIG.json()}))
-@pytest.mark.parametrize("leader", (True, False))
-def test_oci_runtime_ready_event_handler(requirer_ctx, leader, remote_app_data) -> None:
-    """Test that an OCI runtime requirer can consume `oci.conf` data from a runtime provider."""
-    oci_runtime_integration_id = 22
-    oci_runtime_integration = testing.Relation(
-        endpoint=OCI_RUNTIME_INTEGRATION_NAME,
-        interface="oci-runtime",
-        id=oci_runtime_integration_id,
-        remote_app_name="oci-runtime-provider",
-        remote_app_data=remote_app_data,
-    )
-
-    requirer_ctx.run(
-        requirer_ctx.on.relation_changed(oci_runtime_integration),
-        testing.State(leader=leader, relations={oci_runtime_integration}),
-    )
-
-    if leader and remote_app_data:
-        # `MockOCIRunTimeRequirerCharm` provides an assertion to check if `remote_app_data`
-        #  can be read correctly to consume `oci.conf` data.
-
-        # Assert that the last event emitted on the `leader` unit is an `OCIRunTimeReadyEvent`.
-        assert isinstance(requirer_ctx.emitted_events[-1], OCIRuntimeReadyEvent)
-
-        # Assert that `OCIRunTimeReadyEvent` was emitted only once.
-        occurred = defaultdict(lambda: 0)
-        for event in requirer_ctx.emitted_events:
-            occurred[type(event)] += 1
-
-        assert occurred[OCIRuntimeReadyEvent] == 1
-
-    else:
-        # Assert that `OCIRunTimeReadyEvent` is never emitted on non-leader units or on the
-        # leader unit if `remote_app_data` is empty - e.g. blank `RelationChangedEvent` emitted
-        # after `slurmctld` and `oci-runtime-provider` are integrated together.
-        assert not any(
-            isinstance(event, OCIRuntimeReadyEvent) for event in requirer_ctx.emitted_events
+    def test_requirer_oci_runtime_ready_event(self, requirer_ctx, leader, ready) -> None:
+        """Test that an OCI runtime requirer can consume `oci.conf` data from a runtime provider."""
+        oci_runtime_integration_id = 1
+        oci_runtime_integration = testing.Relation(
+            endpoint=OCI_RUNTIME_INTEGRATION_NAME,
+            interface="oci-runtime",
+            id=oci_runtime_integration_id,
+            remote_app_name="oci-runtime-provider",
+            remote_app_data={"ociconfig": EXAMPLE_OCI_CONFIG.json()} if ready else {},
         )
 
-
-@pytest.mark.parametrize("leader", (True, False))
-def test_oci_runtime_disconnected_event_handler(requirer_ctx, leader) -> None:
-    """Test that an OCI requirer properly captures when the runtime provider is disconnected."""
-    oci_runtime_integration_id = 22
-    oci_runtime_integration = testing.Relation(
-        endpoint=OCI_RUNTIME_INTEGRATION_NAME,
-        interface="oci-runtime",
-        id=oci_runtime_integration_id,
-        remote_app_name="oci-runtime-provider",
-    )
-
-    requirer_ctx.run(
-        requirer_ctx.on.relation_broken(oci_runtime_integration),
-        testing.State(leader=leader, relations={oci_runtime_integration}),
-    )
-
-    if leader:
-        # Assert that the last event emitted on the `leader` unit is
-        # an `OCIRunTimeDisconnectedEvent`.
-        assert isinstance(requirer_ctx.emitted_events[-1], OCIRuntimeDisconnectedEvent)
-
-        # Assert that `OCIRunTimeDisconnectedEvent` was emitted only once.
-        occurred = defaultdict(lambda: 0)
-        for event in requirer_ctx.emitted_events:
-            occurred[type(event)] += 1
-
-        assert occurred[OCIRuntimeDisconnectedEvent] == 1
-    else:
-        # Assert that `OCIRunTimeDisconnectEvent` is never emitted on non-leader units.
-        assert not any(
-            isinstance(event, OCIRuntimeDisconnectedEvent) for event in requirer_ctx.emitted_events
+        requirer_ctx.run(
+            requirer_ctx.on.relation_changed(oci_runtime_integration),
+            testing.State(leader=leader, relations={oci_runtime_integration}),
         )
+
+        if leader and ready:
+            # Assert that `OCIRunTimeReadyEvent` was emitted only once.
+            occurred = defaultdict(lambda: 0)
+            for event in requirer_ctx.emitted_events:
+                occurred[type(event)] += 1
+
+            assert occurred[OCIRuntimeReadyEvent] == 1
+
+        else:
+            # Assert that `OCIRunTimeReadyEvent` is never emitted on non-leader units or on the
+            # leader unit if `remote_app_data` is empty - e.g. blank `RelationChangedEvent` emitted
+            # after `slurmctld` and `oci-runtime-provider` are integrated together.
+            assert not any(
+                isinstance(event, OCIRuntimeReadyEvent) for event in requirer_ctx.emitted_events
+            )
+
+    def test_requirer_oci_runtime_disconnected_event(self, requirer_ctx, leader) -> None:
+        """Test that an OCI requirer properly captures when the runtime provider is disconnected."""
+        oci_runtime_integration_id = 1
+        oci_runtime_integration = testing.Relation(
+            endpoint=OCI_RUNTIME_INTEGRATION_NAME,
+            interface="oci-runtime",
+            id=oci_runtime_integration_id,
+            remote_app_name="oci-runtime-provider",
+        )
+
+        requirer_ctx.run(
+            requirer_ctx.on.relation_broken(oci_runtime_integration),
+            testing.State(leader=leader, relations={oci_runtime_integration}),
+        )
+
+        if leader:
+            # Assert that `OCIRunTimeDisconnectedEvent` was emitted only once.
+            occurred = defaultdict(lambda: 0)
+            for event in requirer_ctx.emitted_events:
+                occurred[type(event)] += 1
+
+            assert occurred[OCIRuntimeDisconnectedEvent] == 1
+        else:
+            # Assert that `OCIRunTimeDisconnectEvent` is never emitted on non-leader units.
+            assert not any(
+                isinstance(event, OCIRuntimeDisconnectedEvent)
+                for event in requirer_ctx.emitted_events
+            )
