+++ tests/unit/interfaces/slurmrestd/test_slurmrestd.py
@@ -156,9 +156,6 @@ class TestSlurmrestdInterface:
         )
 
         if ready:
-            # Assert that the last event emitted on all units is a `SlurmctldReadyEvent`.
-            assert isinstance(provider_ctx.emitted_events[-1], SlurmctldReadyEvent)
-
             # Assert that `SlurmctldReadyEvent` was emitted only once.
             occurred = defaultdict(lambda: 0)
             for event in provider_ctx.emitted_events:
